orca_core/hardware/testing_ui/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.6.0
pyserial==3.5
python-socketio>=5.12.0

